+++ services/core-service/src/apps/calendar/commands/impl/update-event.command.ts
@@ -1,0 +1,9 @@
+import { CreateEventRequestDto } from "../../dto/request/create-event.request.dto"
+
+export class UpdateEventCommand {
+  constructor(
+    public readonly userId: string,
+    public readonly eventId: string,
+    public readonly dto: CreateEventRequestDto
+  ) {}
+}