+++ services/core-service/src/apps/calendar/event.controller.ts
@@ -9,13 +9,14 @@ import {
   Body,
   BadRequestException,
   Query,
+  Put,
 } from "@nestjs/common"
 import { EventService } from "./event.service"
 import { statusMessages } from "@/shared/constants/status-messages"
 import { AuthGuard, ModRequest } from "@/auth/auth.guard"
 import { CreateEventRequestDto } from "./dto/request/create-event.request.dto"
 
-@Controller("apps/planner/event")
+@Controller("apps/calendar/event")
 export class EventController {
   constructor(private readonly service: EventService) {}
 
@@ -45,6 +46,37 @@ export class EventController {
         request.user.userId,
         selectedMonth
       )
+    } catch (error) {
+      throw new BadRequestException(
+        error.message || statusMessages.connectionError
+      )
+    }
+  }
+
+  @UseGuards(AuthGuard)
+  @Get(":eventId")
+  async findById(
+    @Request() request: ModRequest,
+    @Param("eventId") eventId: string
+  ) {
+    try {
+      return await this.service.findById(request.user.userId, eventId)
+    } catch (error) {
+      throw new BadRequestException(
+        error.message || statusMessages.connectionError
+      )
+    }
+  }
+
+  @UseGuards(AuthGuard)
+  @Put(":eventId")
+  async updateById(
+    @Request() request: ModRequest,
+    @Param("eventId") eventId: string,
+    @Body() dto: CreateEventRequestDto
+  ) {
+    try {
+      return await this.service.updateById(request.user.userId, eventId, dto)
     } catch (error) {
       throw new BadRequestException(
         error.message || statusMessages.connectionError