+++ services/core-service/src/apps/calendar/queries/impl/find-event-by-id.query.ts
@@ -1,0 +1,6 @@
+export class FindEventByIdQuery {
+  constructor(
+    public readonly userId: string,
+    public readonly eventId: string
+  ) {}
+}